Check integrity of downloaded NetCDF file and retry corrupted files
